SSE-3210 Add network AllowRule to enable product page -> Jira API
SSE-3210 allows the product page registration form to communicate
directly with the Jira API (an atlassian product).

We're experiencing the integration working fine locally but failing
acceptance testing on preview.

We suspect this is due to exposure to network rules. Indeed here we seem
to be allowing sucati allow rules for other major integrations but the
atlassian domain is not allowed.

This adds the root atlassian domain, allowing a 443 HTTPS connection,
following the example of previous it gives it an SID of 2008, my logic
being take the previous highest number and add 1.