LS-2023 Verify sign-in.service.gov.uk with new google DNS code
Woop, worked out why (#36) wasn't working. There was a typo in the
google console value, we were trying to verify ownership of
`sign-in.sevice.gov.uk` which does not exist.

I've generated a new property which generates a new code, so backfilling
that. Should be able to get it verified now.